Test submission
Part 1:
Base AI with returning to the ring and walking up to the target when they were inside the ring. As well as powerup checks and prioritization

Part 2:
Prediction and dodging implemented.

Test submission:
State management improvements, fixed issues with prediction and dodging.

The AI is still rather buggy, so after the test I will try and find some major bugs, and fix them accordingly.